+++ horde/apis/models/kobold_v2.py
@@ -54,17 +54,6 @@ class TextParsers(v2.Parsers):
             help="The seed of the text generation.",
             location="json",
         )
-        self.generate_parser.add_argument(
-            "allow_downgrade",
-            type=bool,
-            default=False,
-            required=False,
-            help=(
-                "If True, The request and the user or shared key doesn't have enough upfront kudos to perform, "
-                "it will be downgraded so that it doesn't require upfront kudos. "
-            ),
-            location="json",
-        )
 
 
 class TextModels(v2.Models):
@@ -332,6 +321,13 @@ class TextModels(v2.Models):
                     description=(
                         "When true, This request will not use batching. This will allow you to retrieve accurate seeds. "
                         "Feature is restricted to Trusted users and Patreons."
+                    ),
+                ),
+                "allow_downgrade": fields.Boolean(
+                    default=False,
+                    description=(
+                        "When true and the request requires upfront kudos and the account does not have enough "
+                        "The request will be downgraded in max context and max tokens so that it does not need upfront kudos."
                     ),
                 ),
                 "webhook": fields.String(