+++ horde/apis/v2/base.py
@@ -196,7 +196,10 @@ class GenerateTemplate(Resource):
                 if len(self.args.extra_source_images) > 5:
                     raise e.BadRequest("You can send a maximum of 5 extra source images.", rc="TooManyExtraSourceImages.")
                 if len(self.args.extra_source_images) > 1 and not self.user.trusted and not patrons.is_patron(self.user.id):
-                    raise e.BadRequest("Only trusted users and patrons can send more than 1 extra source images.", rc="MoreThanMinExtraSourceImage.")
+                    raise e.BadRequest(
+                        "Only trusted users and patrons can send more than 1 extra source images.",
+                        rc="MoreThanMinExtraSourceImage.",
+                    )
             if self.user.education or self.user.trusted:
                 lim.dynamic_ip_whitelist.whitelist_ip(self.user_ip)
             self.username = self.user.get_unique_alias()